requirements.txt
Brotli==1.0.9
click==8.1.3
colorama==0.4.4
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distlib==0.3.4
filelock==3.7.0
Flask==2.1.2
Flask-Compress==1.12
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
platformdirs==2.5.2
six==1.16.0
tenacity==8.0.1
typing-extensions==4.2.0
virtualenv==20.14.1
Werkzeug==2.1.2
zipp==3.8.0
